+++ Smart_Office_Service/src/main/java/grpc/services/hr_department_service/WeeklyWorkingHourResponseOrBuilder.java
@@ -8,9 +8,9 @@ public interface WeeklyWorkingHourResponseOrBuilder extends
     com.google.protobuf.MessageOrBuilder {
 
   /**
-   * <code>int32 total_weekly_working_hour = 1;</code>
+   * <code>double total_weekly_working_hour = 1;</code>
    */
-  int getTotalWeeklyWorkingHour();
+  double getTotalWeeklyWorkingHour();
 
   /**
    * <code>string weekly_working_hour_message = 2;</code>
